Problem 2: use BigInteger instead of int

+++ C#/Problem0002/Problem0002/FibonacciNumbers.cs
@@ -1,12 +1,13 @@
 ﻿using System.Collections.Generic;
+using System.Numerics;
 
 namespace Problem0002
 {
     public class FibonacciNumbers
     {
-        public List<int> GenerateFibonacciSeries(int limitValueTo)
+        public List<BigInteger> GenerateFibonacciSeries(BigInteger limitValueTo)
         {
-            var fibonacciSeries = new List<int> { 1, 1 };
+            var fibonacciSeries = new List<BigInteger> { 1, 1 };
             for (var i = 1; fibonacciSeries[i] < limitValueTo; i++)
             {
                 fibonacciSeries.Add(fibonacciSeries[i] + fibonacciSeries[i - 1]);
